Change StyleCop
