+++ ClimateExplorer.Web.Client/Services/IExporter.cs
@@ -1,0 +1,7 @@
+﻿using ClimateExplorer.Core.Model;
+using ClimateExplorer.Web.UiModel;
+
+public interface IExporter
+{
+    Stream ExportChartData(ILogger logger, IEnumerable<GeographicalEntity> locations, DataDownloadPackage dataDownloadPackage, string sourceUri);
+}
