+++ ClimateExplorer.Web.Client/Services/ISiteOverviewService.cs
@@ -1,0 +1,7 @@
+﻿namespace ClimateExplorer.Web.Client.Services;
+
+public interface ISiteOverviewService
+{
+    event Action? ShowRequested;
+    void RequestShow();
+}
