+++ ClimateExplorer.Web.Client/Services/InfoPanelDismissal.cs
@@ -1,0 +1,8 @@
+﻿namespace ClimateExplorer.Web.Client.Services;
+
+public class InfoPanelDismissal
+{
+    public string Name { get; set; } = string.Empty;
+    public string Version { get; set; } = string.Empty;
+    public DateTime DismissedAtUtc { get; set; }
+}
