+++ ClimateExplorer.Web.Client/Services/InfoPanelDismissalService.cs
@@ -1,12 +1,6 @@
-namespace ClimateExplorer.Web.Services;
+namespace ClimateExplorer.Web.Client.Services;
 
 using Blazored.LocalStorage;
-
-public interface IInfoPanelDismissalService
-{
-    Task<bool> ShouldShowAsync(string panelName, string version);
-    Task DismissAsync(string panelName, string version);
-}
 
 public class InfoPanelDismissalService : IInfoPanelDismissalService
 {
@@ -67,11 +61,4 @@ public class InfoPanelDismissalService : IInfoPanelDismissalService
     {
         return await localStorage.GetItemAsync<List<InfoPanelDismissal>>(StorageKey) ?? [];
     }
-}
-
-public class InfoPanelDismissal
-{
-    public string Name { get; set; } = string.Empty;
-    public string Version { get; set; } = string.Empty;
-    public DateTime DismissedAtUtc { get; set; }
 }
