+++ ClimateExplorer.Web/Program.cs
@@ -5,7 +5,7 @@ using Blazorise.Bootstrap5;
 using Blazorise.Icons.FontAwesome;
 using ClimateExplorer.Web;
 using ClimateExplorer.Web.Client.Pages;
-using ClimateExplorer.Web.Services;
+using ClimateExplorer.Web.Client.Services;
 using ClimateExplorer.WebApiClient.Services;
 using CurrentDevice;
 using DPBlazorMapLibrary;
